Convert everything to structs in their own files
