+++ Sources/DiceKit/Dice.swift
@@ -3,7 +3,7 @@
 /// The properties of `Dice` objects are immutable; use the addition operators to combine them with other `Die` objects or modifiers. You can use compound assignment operators if you want, so long as you declare the `Dice` object as a `var` instead of a `let` constant.
 ///
 /// - Author: Samasaur
-public class Dice {
+public struct Dice {
     /// The dice that make up this collection, along with how many times they appear.
     ///
     /// This `[Die: Int]` dictionary stores the types of dice that appear, paired with the number of times they appear. For example:
@@ -257,16 +257,22 @@ public class Dice {
     public init(copyOf other: Dice) {
         var newDice: [Die: Int] = [:]
         for (d, c) in other.dice {
-            newDice[d.copy()] = c
+            newDice[d] = c
         }
         self.dice = newDice
         modifier = other.modifier
     }
 
+    private let __probabilities = LazyBox<Dice, Chances> { d in
+        return d.calculateChances()
+    }
+
     /// The probabilities of all possible rolls.
     ///
     /// - Since: 0.17.0
-    public lazy var probabilities: Chances = calculateChances()
+    public var probabilities: Chances {
+        return __probabilities.value(input: self)
+    }
 }
 
 extension Dice: Rollable {
@@ -521,7 +527,7 @@ public extension Dice {
         var dice: [Die] = []
         for (d, c) in lhs.dice {
             for _ in 0..<c {
-                dice.append(d.copy())
+                dice.append(d)
             }
         }
         dice.append(rhs)
@@ -537,7 +543,7 @@ public extension Dice {
         var dice: [Die] = []
         for (d, c) in rhs.dice {
             for _ in 0..<c {
-                dice.append(d.copy())
+                dice.append(d)
             }
         }
         dice.append(lhs)
@@ -554,12 +560,12 @@ public extension Dice {
         var dice: [Die] = []
         for (d, c) in lhs.dice {
             for _ in 0..<c {
-                dice.append(d.copy())
+                dice.append(d)
             }
         }
         for (d, c) in rhs.dice {
             for _ in 0..<c {
-                dice.append(d.copy())
+                dice.append(d)
             }
         }
         return Dice(dice: dice, withModifier: rhs.modifier + lhs.modifier)
@@ -575,7 +581,7 @@ public extension Dice {
         var dice: [Die] = []
         for (d, c) in lhs.dice {
             for _ in 0..<c {
-                dice.append(d.copy())
+                dice.append(d)
             }
         }
         return Dice(dice: dice, withModifier: lhs.modifier + rhs)
@@ -590,7 +596,7 @@ public extension Dice {
         var dice: [Die] = []
         for (d, c) in rhs.dice {
             for _ in 0..<c {
-                dice.append(d.copy())
+                dice.append(d)
             }
         }
         return Dice(dice: dice, withModifier: lhs + rhs.modifier)
