+++ Sources/DiceKit/IncludableInDice.swift
@@ -1,0 +1,3 @@
+public protocol IncludableInDice: Rollable {
+    var doubleAverageResult: Double { get }
+}
