+++ Sources/DiceKit/Typealiases.swift
@@ -1,0 +1,22 @@
+/// See `Die`.
+public typealias DKDie = Die
+/// See `Dice`.
+public typealias DKDice = Dice
+/// See `Roll`.
+public typealias DKRoll = Roll
+/// See `Rollable`.
+public typealias DKRollable = Rollable
+/// See `MultipleRollResult`.
+public typealias DKMultipleRollResult = MultipleRollResult
+/// The result of any `roll` method from any `Rollable` type.
+///
+/// - Since: 0.8.0
+public typealias Roll = Int
+/// See `Error`.
+public typealias DKError = DiceKit.Error
+/// See `WeightedDie`.
+public typealias DKWeightedDie = WeightedDie
+/// See `Chance`.
+public typealias DKChance = Chance
+/// See `Chances`.
+public typealias DKChances = Chances
